expose SSRManifest
