+++ packages/kit/types/app.d.ts
@@ -1,6 +1,6 @@
 import { ReadOnlyFormData, RequestHeaders } from './helper';
 import { ServerResponse } from './hooks';
-import { PrerenderOptions, SSRManifest } from './internal';
+import { PrerenderOptions, SSRNodeLoader, SSRRoute } from './internal';
 
 export class App {
 	constructor(manifest: SSRManifest);
@@ -28,4 +28,20 @@ export interface IncomingRequest {
 	query: URLSearchParams;
 	headers: RequestHeaders;
 	rawBody: RawBody;
+}
+
+export interface SSRManifest {
+	appDir: string;
+	assets: Set<string>;
+	/** private fields */
+	_: {
+		mime: Record<string, string>;
+		entry: {
+			file: string;
+			js: string[];
+			css: string[];
+		};
+		nodes: SSRNodeLoader[];
+		routes: SSRRoute[];
+	};
 }
