+++ packages/kit/types/index.d.ts
@@ -3,7 +3,7 @@
 
 import './ambient-modules';
 
-export { App, IncomingRequest, RawBody } from './app';
+export { App, IncomingRequest, RawBody, SSRManifest } from './app';
 export { Adapter, Builder, Config, PrerenderErrorHandler, ValidatedConfig } from './config';
 export { EndpointOutput, RequestHandler } from './endpoint';
 export { ErrorLoad, ErrorLoadInput, Load, LoadInput, LoadOutput, Page } from './page';
